requirements.txt
botocore==1.5.45
argparse
PyYaml

